requirements.txt
Django==1.5.2
South==0.8.2
amqp==1.0.13
anyjson==0.3.3
billiard==2.7.3.32
celery==3.0.22
distribute==0.6.49
django-celery==3.0.21
django-debug-toolbar==0.9.4
django-model-utils==1.4.0
kombu==2.5.13
python-dateutil==2.1
pytz==2013b
six==1.3.0
unicodecsv==0.9.4
wsgiref==0.1.2
